+++ fficxx/lib/FFICXX/Generate/ContentMaker.hs
@@ -30,12 +30,12 @@ import           Language.Haskell.Exts.Syntax           (Module(..),Decl(..))
 import           System.FilePath
 --
 import           FFICXX.Generate.Code.Cpp
-import           FFICXX.Generate.Code.Dependency
 import           FFICXX.Generate.Code.HsFFI             (genHsFFI
                                                         ,genImportInFFI
                                                         ,genTopLevelFuncFFI)
 import           FFICXX.Generate.Code.HsFrontEnd
 import           FFICXX.Generate.Code.Primitive
+import           FFICXX.Generate.Dependency
 import           FFICXX.Generate.Type.Annotate
 import           FFICXX.Generate.Type.Class
 import           FFICXX.Generate.Type.Module
